+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Accept `Update(Person)` C2S activities.
+- Disable same-origin check if fetched object is portable.
 
 ## [2.23.0] - 2024-06-23